% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Somar.R
\name{somar}
\alias{somar}
\title{Função para somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{um número a ser somado}

\item{b}{outro número a ser somado}
}
\description{
Esta função serve para somar dois números
}
\examples{
somar (1,2)
resposta 3


}
