% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Função para subtrair}
\usage{
subtrair(a, b)
}
\arguments{
\item{a}{um número a ser subtraido}

\item{b}{outro número a subtrair}
}
\description{
Esta função serve para subtrair dois números
}
\examples{
subtrair (4,2)
resposta 2


}
